+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -8,6 +8,7 @@ import RicosModal from './modals/RicosModal';
 import './styles.css';
 import { RicosEditorProps, EditorDataInstance, RichContentChild } from './index';
 import { hasActiveUploads } from './utils/hasActiveUploads';
+import { convertToRaw } from 'wix-rich-content-editor/dist/lib/editorStateConversion';
 
 const filterDraftEditorSettings = (draftEditorSettings: Partial<EditorProps>) =>
   Object.entries(draftEditorSettings).map(
@@ -84,10 +85,12 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
   };
 
   onBusyChange = (contentState: ContentState) => {
+    const { onBusyChange, onChange } = this.props;
     const isBusy = hasActiveUploads(contentState);
     if (this.isBusy !== isBusy) {
       this.isBusy = isBusy;
-      this.props.onBusyChange?.(isBusy);
+      onBusyChange?.(isBusy);
+      onChange?.(convertToRaw(contentState));
     }
   };