+++ CHANGELOG.md
@@ -20,7 +20,9 @@
 ### :rocket: New Feature
 - `image`
   - [#1408](https://github.com/wix-incubator/rich-content/pull/1408) upload error displays in image settings
-
+### :bug: Bug Fix
+- `ricos-editor`
+  - [#1476](https://github.com/wix-incubator/rich-content/pull/1476) onChange is called when upload starts/ends
 ### 🏠 Internal
 - `general`
   - [#1408](https://github.com/wix-incubator/rich-content/pull/1408) media upload errors saved and handled from component data
@@ -31,7 +33,7 @@
 ## 7.16.14 (Aug 23, 2020)
 ### :bug: Bug Fix
 - `ricos-editor`
-  - [#1471](https://github.com/wix-incubator/rich-content/pull/1471/) missing height in ricos editor wrapping div
+  - [#1471](https://github.com/wix-incubator/rich-content/pull/1471) missing height in ricos editor wrapping div
 - `editor-common`
   - [#1473](https://github.com/wix-incubator/rich-content/pull/1473) fix RadioGroup not working in html plugin